examples implementation
outputs different result as data on each example selected. only 1 and 2 implemented so far. adding a query to the commented lines on server.js will make the others work too. (data all clustered together, didnt bother organzing since we're gonna put it in a graph anyway, but u can inspect element and go to console to see a formatted view of the data)